refactor(cli): extract lookup and runtime tester commands
