+++ .changeset/cli-command-islands.md
@@ -1,0 +1,5 @@
+---
+"skillset": patch
+---
+
+Extract lookup and runtime tester CLI command islands from the shared CLI core.
